Fix bug with `pip list` parsing
