+++ src/test/constants.ts
@@ -1,0 +1,4 @@
+import * as path from 'path';
+
+export const EXTENSION_ROOT = path.dirname(path.dirname(__dirname));
+export const EXTENSION_TEST_ROOT = path.join(EXTENSION_ROOT, 'src', 'test');
