+++ src/test/managers/builtin/piplist1.expected.json
@@ -1,0 +1,37 @@
+{
+    "packages": [
+        { "name": "argcomplete", "version": "3.1.2" },
+        { "name": "black", "version": "23.12.1" },
+        { "name": "build", "version": "1.2.1" },
+        { "name": "click", "version": "8.1.7" },
+        { "name": "colorama", "version": "0.4.6" },
+        { "name": "colorlog", "version": "6.7.0" },
+        { "name": "coverage", "version": "7.6.1" },
+        { "name": "distlib", "version": "0.3.7" },
+        { "name": "exceptiongroup", "version": "1.1.3" },
+        { "name": "filelock", "version": "3.13.1" },
+        { "name": "importlib_metadata", "version": "7.1.0" },
+        { "name": "iniconfig", "version": "2.0.0" },
+        { "name": "isort", "version": "5.13.2" },
+        { "name": "mypy-extensions", "version": "1.0.0" },
+        { "name": "namedpipe", "version": "0.1.1" },
+        { "name": "nox", "version": "2024.3.2" },
+        { "name": "packaging", "version": "23.2" },
+        { "name": "pathspec", "version": "0.12.1" },
+        { "name": "pip", "version": "24.0" },
+        { "name": "pip-tools", "version": "7.4.1" },
+        { "name": "platformdirs", "version": "3.11.0" },
+        { "name": "pluggy", "version": "1.4.0" },
+        { "name": "pyproject_hooks", "version": "1.1.0" },
+        { "name": "pytest", "version": "8.1.1" },
+        { "name": "pytest-cov", "version": "5.0.0" },
+        { "name": "pywin32", "version": "306" },
+        { "name": "ruff", "version": "0.7.4" },
+        { "name": "setuptools", "version": "56.0.0" },
+        { "name": "tomli", "version": "2.0.1" },
+        { "name": "typing_extensions", "version": "4.9.0" },
+        { "name": "virtualenv", "version": "20.24.6" },
+        { "name": "wheel", "version": "0.43.0" },
+        { "name": "zipp", "version": "3.19.2" }
+    ]
+}
