+++ src/test/managers/builtin/piplist2.actual.txt
@@ -1,0 +1,35 @@
+Package            Version
+------------------ --------
+argcomplete        3.1.2
+black              23.12.1
+build              1.2.1
+click              8.1.7
+colorama           0.4.6
+colorlog           6.7.0
+coverage           7.6.1
+distlib            0.3.7
+exceptiongroup     1.1.3
+filelock           3.13.1
+importlib_metadata 7.1.0
+iniconfig          2.0.0
+isort              5.13.2
+mypy-extensions    1.0.0
+namedpipe          0.1.1
+nox                2024.3.2
+packaging          23.2
+pathspec           0.12.1
+pip                24.0
+pip-tools          7.4.1
+platformdirs       3.11.0
+pluggy             1.4.0
+pyproject_hooks    1.1.0
+pytest             8.1.1
+pytest-cov         5.0.0
+pywin32            306
+ruff               0.7.4
+setuptools         56.0.0
+tomli              2.0.1
+typing_extensions  4.9.0
+virtualenv         20.24.6
+wheel              0.43.0
+zipp               3.19.2
